% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlotLims.R
\name{getPlotLims}
\alias{getPlotLims}
\title{\code{getPlotLims} Get the plot limits of a volcano plot for the passed DESeqResults object}
\usage{
getPlotLims(DDSresult)
}
\arguments{
\item{DDSresult}{a DESeqResults object}
}
\value{
list of xlim and ylim of a vplot for the object
}
\description{
\code{getPlotLims} Get the plot limits of a volcano plot for the passed DESeqResults object
}
\examples{
require(DESeq2)
simDDS <- DESeq(simDDS)
res <- results(simDDS)
getPlotLims(res)

}
\seealso{
\link{par}
}
