% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{\code{overlay} create a layer on an existing volcano plot}
\usage{
overlay(DDSresult)
}
\arguments{
\item{DDSresult}{a DESeqResults object}
}
\description{
\code{overlay} should be executed after vplot(), and can be re-executed several times to build up layers on a plot.
}
\examples{
load(system.file("extdata/dev/exampleDDSresult.RData", package = "pointszr"))
vplot(exDDSresult, main = "Example output for vplot()")



}
