% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPointszrApp.R
\name{runPointszrApp}
\alias{runPointszrApp}
\title{\code{runPointszrApp} launch the shiny app distributed with this package framework}
\usage{
runPointszrApp()
}
\value{
no return, app side effect
}
\description{
\code{runPointszrApp} launches the shiny app for which the code has been \cr
placed in \code{./inst/shiny-scripts/pointszrApp/}. You can also find it \cr
online at https://harrig12.shinyapps.io/pointszrApp/
}
\examples{
\dontrun{
runPointszrApp()
}
}
