% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/szPreview.R
\name{szPreview}
\alias{szPreview}
\title{\code{szPreview} Display what a list of point sizes looks like when plotted}
\usage{
szPreview(sz)
}
\arguments{
\item{sz}{an list of sizes to make points}
}
\value{
no return, plotting side effect
}
\description{
\code{szPreview} Display what a list of point sizes looks like when plotted
}
\examples{
sizes <- lseq(0.5, 8, 6)
szPreview(sizes)
}
\seealso{
\link{par}
}
