% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.R
\name{vplot}
\alias{vplot}
\title{\code{vplot} create a volcano plot}
\usage{
vplot(DDSresult, main = "Volcano Plot", xlab = "log2 Fold Change",
  ylab = "-log10 p-value", pch = 20, cex = 0.5, ...)
}
\arguments{
\item{DDSresult}{a DESeqResults object}

\item{main}{plot title}

\item{xlab}{plot x axis label}

\item{ylab}{plot y axis label}

\item{pch}{point style to be displayed. by default pch = 20.}

\item{cex}{point size to be displayed. by default cex = 0.5}

\item{...}{additional arguments to pass to plot().}
}
\value{
no return, plotting side effect
}
\description{
\code{vplot} plot the differentially expressed genes, as specified in DDSresult
}
\examples{
require(DESeq2)
simDDS <- DESeq(simDDS)
res <- results(simDDS)
vplot(res, xlab = "log2FoldChange", ylab = "-log10(pvalue)")

}
\seealso{
\link{par}
}
