% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.R
\name{vplot}
\alias{vplot}
\title{\code{vplot} create a volcano plot}
\usage{
vplot(DDSresult, szModifier, method, colSig = T, ...)
}
\arguments{
\item{DDSresult}{a DESeqResults object}

\item{szModifier}{a vector containing the variable that point size is varied based on}

\item{method}{a string that selects the method to use when scaling point size}

\item{colSig}{colour points based on significance. Red if padj<0.05, orange of log2FC>1, green if both. (from http://www.gettinggeneticsdone.com/2014/05/r-volcano-plots-to-visualize-rnaseq-microarray.html)}

\item{...}{arguments to pass to plot()}
}
\value{
returns the DESeqResults object that was used for plotting.
}
\description{
\code{vplot}
}
\examples{
load(system.file("extdata/dev/exampleDDSresult.RData", package = "pointszr"))
vplot(exDDSresult, main = "Example output for vplot()")
}
