% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.R
\name{vplot}
\alias{vplot}
\title{\code{vplot} create a volcano plot}
\usage{
vplot(DDSresult, pch = 20, cex = 0.5, ...)
}
\arguments{
\item{DDSresult}{a DESeqResults object}

\item{pch}{point style to be displayed. by default pch = 20.}

\item{cex}{point size to be displayed. by default cex = 0.5}

\item{...}{additional arguments to pass to plot().}
}
\value{
returns the DESeqResults object that was used for plotting.
}
\description{
\code{vplot}
}
\examples{
library(DESeq2)
simDDS <- DESeq(simDDS)
res <- results(simDDS)

}
\seealso{
DESeq2
}
